@@ -67,29 +70,29 @@
     "polished": "^3.3.2",
     "prettier": "^2.2.1",
     "qs": "^6.9.4",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-confetti": "^6.0.0",
     "react-device-detect": "^1.6.2",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-feather": "^2.0.8",
     "react-ga": "^2.5.7",
     "react-i18next": "^10.7.0",
     "react-markdown": "^4.3.1",
     "react-popper": "^2.2.3",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "^4.0.2",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "react-spring": "^8.0.27",
     "react-use-gesture": "^6.0.14",
     "react-virtualized-auto-sizer": "^1.0.2",
     "react-window": "^1.8.5",
     "rebass": "^4.0.7",
     "redux-localstorage-simple": "^2.3.1",
-    "serve": "^11.3.0",
+    "serve": "^11.3.2",
     "start-server-and-test": "^1.11.0",
     "styled-components": "^4.2.0",
     "styled-system": "^5.1.5",
-    "typescript": "^3.8.3",
+    "typescript": "^4.4.3",
     "use-count-up": "^2.2.5",
     "wcag-contrast": "^3.0.0"
   },
