@@ -1,17 +1,12 @@
 import React from 'react'
-import { Redirect, RouteComponentProps } from 'react-router-dom'
+import { Navigate, useParams } from 'react-router-dom'
 import TokenPage from './TokenPage'
 import { isAddress } from 'ethers/lib/utils'
 
-export function RedirectInvalidToken(props: RouteComponentProps<{ address: string }>) {
-  const {
-    match: {
-      params: { address },
-    },
-  } = props
-
-  if (!isAddress(address)) {
-    return <Redirect to={`/`} />
+export function RedirectInvalidToken() {
+  const { address } = useParams()
+  if (address && !isAddress(address)) {
+    return <Navigate to={`/`} />
   }
-  return <TokenPage {...props} />
+  return <TokenPage />
 }
