@@ -154,7 +154,7 @@ function StatusIcon({ connector }: { connector: AbstractConnector }) {
 
 function Web3StatusInner() {
   const { t } = useTranslation()
-  const { account, connector, error } = useWeb3React()
+  const { account, connector } = useWeb3React()
 
   const { ENSName } = useENSName(account ?? undefined)
 
