@@ -3,12 +3,12 @@
  *
  */
 
-import { OpUnitType } from 'dayjs'
+import { ManipulateType } from 'dayjs'
 
 export const ONE_HOUR_SECONDS = 3600
 
 export const TimeWindow: {
-  [key: string]: OpUnitType
+  [key: string]: ManipulateType
 } = {
   DAY: 'day',
   WEEK: 'week',
